D++ slash command
Added a commented out D++ API slash command at the bottom of the driver class in order to format the official battle sequence around taking in user commands. The actual project folder necessary to run the bot as an actual bot is too large to be constantly switching between devices, so slash commands are commented out temporarily and will be debugged on the main computer containing the D++ API files.